# public/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Online Shopper",
        "contact": {
            "name": "E4 Class G17 Group ?",
            "url": "https://www.kuylim.wordpress.com"
        },
        "version": "1.0"
    },
    "host": "online-shop.com",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Return all categories",
                "operationId": "findAllCategories",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Category Not Found"
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "description": "Add a new category.",
                "operationId": "addCategory",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "category",
                        "in": "body",
                        "description": "Category to add to the Database",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Save Category Success"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "description": "Returns a categories based on a single ID",
                "operationId": "findCategoryById",
                "produces": [
                    "application/json",
                    "application/xml",
                    "text/html",
                    "text/xml"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category to fetch",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "category found"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "description": "Update old category.",
                "operationId": "updateCategory",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category to update",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "category",
                        "in": "body",
                        "description": "Category to update to the database",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Save Category Success"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "description": "deletes a single category based on the category id",
                "operationId": "deleteCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "category deleted success"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/categories/search/{name}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "description": "Returns a category based on a name",
                "operationId": "findCategoryByName",
                "produces": [
                    "application/json",
                    "application/xml",
                    "text/html",
                    "text/xml"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "name of category to fetch",
                        "required": true,
                        "type": "String",
                        "format": "String"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "category found"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
...
    },
...
}